完善实体类

+++ src/main/java/cn/anselyuki/controller/ProductController.java
@@ -1,5 +1,6 @@
 package cn.anselyuki.controller;
 
+import cn.anselyuki.controller.request.ProductDTO;
 import cn.anselyuki.controller.response.Result;
 import cn.anselyuki.repository.ProductRepository;
 import cn.anselyuki.repository.entity.Product;
@@ -20,19 +21,28 @@ public class ProductController {
 
     @PutMapping("add")
     @Operation(summary = "添加物资", description = "添加物资")
-    public ResponseEntity<Result<Product>> addProduct() {
-        return null;
+    public ResponseEntity<Result<Product>> addProduct(ProductDTO productDTO) {
+        Product product = new Product(productDTO);
+        try {
+            productRepository.save(product);
+        } catch (Exception e) {
+            Result.fail(403, "添加物资失败");
+        }
+        return Result.success(product);
     }
 
     @DeleteMapping("delete/{id}")
     @Operation(summary = "删除物资", description = "删除物资")
     public ResponseEntity<Result<Product>> deleteProduct(@PathVariable String id) {
+        if (!productRepository.existsById(id)) {
+            return Result.fail(404, "物资不存在");
+        }
         try {
             productRepository.deleteById(id);
         } catch (Exception e) {
-            e.printStackTrace();
+            Result.fail(403, "删除物资失败");
         }
-        return null;
+        return Result.success(null);
     }
 
     @PutMapping("update")
